requirements.txt
streamlit
openai-whisper
google-genai
python-docx
pandas
python-dotenv
